新規登録画面、ログイン画面Google修正

+++ database/05_functions.sql
@@ -88,9 +88,9 @@ RETURNS TABLE(available_jpy bigint, pending_jpy bigint) AS $$
 BEGIN
     RETURN QUERY
     SELECT 
-        COALESCE(SUM(CASE WHEN entry_type IN ('seller_net', 'adjustment') THEN amount_jpy ELSE 0 END), 0) - 
-        COALESCE(SUM(CASE WHEN entry_type = 'payout' THEN amount_jpy ELSE 0 END), 0) as available_jpy,
-        COALESCE(SUM(CASE WHEN entry_type = 'seller_net' AND created_at >= CURRENT_DATE - INTERVAL '30 days' THEN amount_jpy ELSE 0 END), 0) as pending_jpy
+        (COALESCE(SUM(CASE WHEN entry_type IN ('seller_net', 'adjustment') THEN amount_jpy ELSE 0 END), 0)::bigint - 
+        COALESCE(SUM(CASE WHEN entry_type = 'payout' THEN amount_jpy ELSE 0 END), 0)::bigint) as available_jpy,
+        COALESCE(SUM(CASE WHEN entry_type = 'seller_net' AND created_at >= CURRENT_DATE - INTERVAL '30 days' THEN amount_jpy ELSE 0 END), 0)::bigint as pending_jpy
     FROM public.ledger_entries
     WHERE seller_id = seller_uuid;
 END;
